moving any wrappers to separate file + organizing code
