+++ apps/test_streaming_scenario.cpp
@@ -356,7 +356,7 @@ int main(int argc, char **argv)
                                 .with_num_pq_chunks(0)
                                 .is_pq_dist_build(false)
                                 .with_tag_type("uint32")
-                                .with_label_type("uint32")
+                                .with_label_type("uint")
                                 .with_data_type(data_type)
                                 .with_index_write_params(params)
                                 .with_data_load_store_strategy(diskann::MEMORY)
