+++ include/index_config.h
@@ -3,7 +3,6 @@
 
 namespace diskann
 {
-
 enum LoadStoreStrategy
 {
     DISK,
