+++ include/types.h
@@ -6,6 +6,7 @@
 #include <cstdint>
 #include <cstddef>
 #include <any>
+#include "any_wrappers.h"
 
 namespace diskann
 {
@@ -14,5 +15,7 @@ typedef uint32_t location_t;
 using DataType = std::any;
 using TagType = std::any;
 using LabelType = std::any;
+using TagVector = AnyWrapper::AnyVector;
+using TagRobinSet = AnyWrapper::AnyRobinSet;
 
 } // namespace diskann
