--- src/abstract_index.cpp
@@ -1,7 +1,0 @@
-#include "common_includes.h"
-#include "abstract_index.h"
-
-namespace diskann
-{
-
-}
