+++ src/index_factory.cpp
@@ -5,7 +5,7 @@ namespace diskann
 
 IndexFactory::IndexFactory(IndexConfig &config) : _config(config)
 {
-    checkConfig();
+    check_config();
 }
 
 std::shared_ptr<AbstractIndex> IndexFactory::instance()
@@ -59,7 +59,7 @@ std::shared_ptr<AbstractIndex> IndexFactory::instance()
     }
 }
 
-void IndexFactory::checkConfig()
+void IndexFactory::check_config()
 {
     if (_config.dynamic_index && !_config.enable_tags)
     {
@@ -84,6 +84,21 @@ void IndexFactory::checkConfig()
         throw ANNException("ERROR: invalid data type : + " + _config.data_type +
                                " is not supported. please select from [float, int8, uint8]",
                            -1);
+    }
+
+    if (!_config.data_strategy)
+    {
+        throw ANNException("Error: data load store stratagy is required to initialize index", -1);
+    }
+
+    if (!_config.metric)
+    {
+        throw ANNException("Error: please pass metric in index config, can't initialize index without a valic metric.",
+                           -1);
+    }
+    if (_config.dimension == 0)
+    {
+        throw ANNException("Error: dimension of an index can not be zero.", -1);
     }
 }
